Changed REST API classes to resources/ package
