+++ src/test/sample_test.py
@@ -1,14 +1,14 @@
-from ..main import myApp
+from ..main.resources.SimpleAPI import Hello, GoodBye
 
 class TestMyApp(object):
     def test_hello(self):
-        h = myApp.Hello()
+        h = Hello()
         assert h.get() == "Hello"
 
     def test_goodbye_get(self):
-        g = myApp.GoodBye()
+        g = GoodBye()
         assert g.get() == "Good Bye"
 
     def test_goodbye_post(self):
-        g = myApp.GoodBye()
+        g = GoodBye()
         assert g.post() == {'good':'bye'}
